+++ src/components/text-input/TextInput.snap.ts
@@ -1,0 +1,14 @@
+// Jest Snapshot v1, https://goo.gl/fbAQLP
+
+exports[`wvui-text-input matches the snapshot 1`] = `
+<div
+  class="wvui-text-input"
+>
+  <input
+    aria-autocomplete="list"
+    autocomplete="off"
+    class="wvui-text-input-field wvui-text-input-field--enabled wvui-text-input-field--search"
+    dir="auto"
+  />
+</div>
+`;