+++ src/entries/wvui.ts
@@ -1,5 +1,7 @@
 import WvuiButton from '@/components/button/Button.vue';
+import WvuiTextInput from '@/components/text-input/TextInput.vue';
 
 export default {
-	WvuiButton
+	WvuiButton,
+	WvuiTextInput
 };